Extraer modal de error de mark.js
Sexto paso de la descomposición de mark.js: showErrorModal(),
closeErrorModalHandler() y el cableado de sus listeners globales
(botón Reintentar, botón Recargar, teclado) se mueven a
mark/error-modal.js.

A diferencia de los modales anteriores, este comparte estado con el
resto del flujo de identificación/cámara: errorModalVisible pausa el
drawLoop y el dwell mientras el modal está abierto, y el flujo de GPS
en segundo plano lo escribe directamente durante el reintento. El
módulo expone isErrorModalVisible()/setErrorModalVisible() para esos
3 puntos de mark.js que lo leen/escriben fuera del modal en sí.

El único cross-call que tenía showErrorModal() (cancelAutoIdentifyDwell)
y el fallback del botón Reintentar sin onRetry explícito (feedback de
cooldown + reset de faceInFrameState) ahora se pasan como callbacks
onShow/onRetryFallback a initErrorModal(), llamado una sola vez al
cablear los listeners. Los 13 call sites de showErrorModal() en
mark.js no cambian. Refactor puro, sin cambios de comportamiento.
